### 2018-05-15 #736 @Kevin-Mattheus-Moerman
@wouterpotters are you interested in reviewing this for [JOSS](http://joss.theoj.org/about)?
