### 2018-05-16 #736 @Kevin-Mattheus-Moerman
It seems like this could be interesting to the simvascular folks. @osmsc @sshadden @jmerkow @alexkaiser @updega2 @gmaher @fkong7 are you interested in reviewing this [Journal of Open Source Software](http://joss.theoj.org/about) submission?
The review process is focused largely on the software and a short paper, see the [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines).

[--> Check software repository :floppy_disk: : <--](https://github.com/vmtk/vmtk)

Here is an example of recent review process:
https://github.com/openjournals/joss-reviews/issues/691
