### 2018-05-18 #736 @wouterpotters
@Kevin-Mattheus-Moerman As this is my first review for JOSS, I have two questions that might be relevant already for the pre-review process:
1. I noticed some of the authors already published [this paper](https://www.ncbi.nlm.nih.gov/pubmed/19002516). This 2008 paper essentially describes all the main features, the structure and the goals of the current toolbox, the open source toolbox is also mentioned. Despite changes to the toolbox over the years one could view this as a duplicate publication; what is the JOSS policy on this? I would like to emphasise that the authors do mention the [previous publication](https://www.ncbi.nlm.nih.gov/pubmed/19002516) explicitly in their manuscript. Furthermore, I also noted that in the past years there were a lot of improvements, but these focus mainly on user interface improvements, relatively small changes and bug fixes. Please correct me if I'm missing key additions since 2008 @rlizzo
2. Back in 2016 and 2017 I committed some small improvements for this repository. (without directly cooperating with the toolbox owners); I guess this is not considered a conflict of interest by JOSS? I have no affiliation with the authors of the manuscript.
