### 2018-05-19 #736 @rlizzo
Thanks for the feedback @wouterpotters,

**In reference to point 1:**

- While the core functionality of the library (outlined in the 2008 paper mentioned) is largely intact, there have been major improvements to functionality, performance, and user-interface. From the [v0.9 version](https://github.com/vmtk/vmtk/tree/v0.9.0) (released in 2010), the improvements have been largely documented on the [vmtk news webpage](http://www.vmtk.org/news/). A quick count shows that there have been 52 new functions added through new `vmtkScripts`. In addition, nearly every script has seen some modifications and improvement through that time. While the list of additions is too numerous to add here, I've highlighted some excerpts from release notes below:

[Particle Tracing](http://www.vmtk.org/news/2013/12/13/new-tutorial-particle-tracing.html)

> vmtkmeshmergetimesteps: a script used to create a mesh file which will contains all the data relative to the velocity components for each timestep we want to generate traces.
vmtkparticletracer: a script used to generate traces (VtkPolydata) from a velocity vector field.
vmtkpathlineanimator: a script used to animate the traces previously generated.

[VMTK 1.2 Release](http://www.vmtk.org/news/2014/01/24/vmtk-1.2-released.html)

> vmtkimagemorphology: a script used to apply binary or grayscale morphology filter to the input image.
vmtkimageotsuthresholds: a script used to automatically label an image using multiple Otsu thresholding.
vmtksurfacemassproperties: a script for clipping using polyball and tube functions.
vmtksurfacearrayoperation: a script that allow curvature-dependent density meshing.
vmtksurfacebooleanoperation: a script used to perform a boolean operation between two surfaces.
vmtksurfaceregiondrawing: a script used to draw a closed contour on a surface and generate a distance field on the surface.
vmtksurfacecurvature: a script that allow to compute curvature of an input surface, optionally reporting results back to a reference surface (by point id).
vmtksurfacepolyballevaluation: a script used to evaluate the polyball function on the vertices of a surface.
vmtksurfaceregiondrawing: a script that allow to draw a closed contour on a surface and generate a distance field on the surface.
vmtkcenterlineinterpolation : a script used to smooth centerlines with a moving average.
vmtkmesharrayoperation: a script that allow to perform an operation between arrays of two meshes and store result in the first mesh.
vmtkmeshconnectivity: a script used to extract the largest connected region or the closest point-connected region from a mesh.
vmtkmeshpolyballevaluation: a script used to evaluate the polyball function on the vertices of a mesh.
vmtkmeshvolume: a script that allow to computes the volume of a mesh.
vmtkmeshcutter: a script used to generate a slice from a mesh, contributed by Simone Manini, Orobix Srl.
vmtkmeshextractpointdata:a script used to extract extract the magnitude value (and its components) of an array for a point id of a mesh, contributed by Simone Manini, Orobix Srl.
vmtkmeshmergetimesteps: a script used to create a mesh file which will contains all the data relative to the velocity components for each timestep we want to generate traces, contributed by Simone Manini, Orobix Srl.
vmtkparticletracer: a script used to generate traces (VtkPolydata) from a velocity vector field, contributed by Simone Manini, Orobix Srl. Check how it works on the particle tracing tutorial
vmtkpathlineanimator: a script used to animate the traces previously generated, contributed by Simone Manini, Orobix Srl. Check how it works on the particle tracing tutorial
Thanks to Martin Sandve Alnaes from Simula, Norway for contributing improvements to capping. Now we can use a mix of methods for different boundaries without losing boundary markers for subsequent meshing.
There is a new algorithm which allows to generate multiple boundary layers of different thickness. Check how it works on the mesh generation tutorial.

[VMTK 1.3 Release](http://www.vmtk.org/news/2016/01/22/vmtk-1.3-released.html)

> Added support for VTK 6.x and VTK 7.x. vmtk now builds against the latest official versions of both VTK and ITK.
vmtkimagebinarize: binarize an input image using two thresholds and producing a label map.
vmtksurfacearraysmoothing: perform smoothing on the values of a point data array defined on a surface.
vmtksurfacecliploop: clip an input surface using a user-defined loop.
vmtksurfaceendclipper: remove branches of a tubular surface by specifying seeds on the surface and using automatically oriented planes.
vmtksurfaceloopextraction: interactively define loops on a surface.

[VMTK 1.4 Release](http://www.vmtk.org/news/2018/03/19/vmtk-1.4.0-released.html)

>We now allow you to use all of the functionality that VMTK offers in filtering images, creating surfaces, centerlines, or volumetric meshes, while seemlessly being able to employ any external processing available in the scientific python ecosystem. With the following new methods, you can losslesly translate data (generated in either VMTK or in Python proper) back and forth between the VTK data formats and Numpy Arrays:
vmtkImageToNumpy
vmtkNumpyToImage
vmtkCenterlinesToNumpy
vmtkNumpyToCenterlines
vmtkSurfaceToNumpy
vmtkNumpyToSurface
vmtkMeshToNumpy
vmtkNumpyToMesh
A tutorial has been provided on the VMTK website which walks you through the process of interacting with the translation routines
With the VMTK-Numpy conversion utilities, we now allow VMTK data structures to be saved (and read) to disk as HDF5 or Python Pickle objects. These methods are available in the following methods:
vmtkNumpyWriter
vmtkNumpyReader

- We have completely revised our build system, and created a whole CI suite / testing / distribution system in order to help us ensure continuity between changes in the code base.
- We've added many new tutorials and documentation updates.
- VMTK extensions for 3D Slicer and Blender have been released.
- @lantiga may be able to elaborate more on this topic, as he is the original author of the toolkit.

**In reference to point 2:**

I just wanted to post the two pull requests which have you tagged in them. Please let us know the policy on this:
- "Added opacity function for vmtkimagemipviewer.py" - https://github.com/vmtk/vmtk/pull/173
- "Added functionality to change the position of a window" - https://github.com/vmtk/vmtk/pull/168
