### 2018-05-19 9ea125d
added utility to allow for cycling of volume property presets by pressing the w key. removed MR DTI preset due to errors it generated
