### 2018-05-19 ad555fa
added multiple colliding fronts method to vmtkimageinitialization. added selectable seed color to vmtkimageseeder
