+++ vmtkScripts/vmtkimageinitialization.py
@@ -18,6 +18,7 @@ import sys
 import math
 import string
 import vtk
+from random import random
 
 from vmtk import vtkvmtk
 
@@ -50,6 +51,7 @@ class vmtkImageInitialization(pypes.pypeScript):
         self.TargetPoints = []
 
         self.NegateImage = 0
+        self.NumSeedsSelected = 0
 
         self.IsoSurfaceValue = 0.0
 
@@ -62,7 +64,7 @@ class vmtkImageInitialization(pypes.pypeScript):
             ['Image','i','vtkImageData',1,'','','vmtkimagereader'],
             ['NegateImage','negate','bool',1,'','negate image values before initializing'],
             ['Interactive','interactive','bool',1,'',''],
-            ['Method','method','str',1,'["isosurface","threshold","collidingfronts","fastmarching","seeds"]',''],
+            ['Method','method','str',1,'["isosurface","threshold","collidingfronts","fastmarching","seeds","multiplecollidingfronts"]',''],
             ['SourcePoints','sourcepoints','int',-1,'','list of source point IJK coordinates'],
             ['TargetPoints','targetpoints','int',-1,'','list of target point IJK coordinates'],
             ['UpperThreshold','upperthreshold','float',1,'','the value of the upper threshold to use for threshold, collidingfronts and fastmarching'],
@@ -75,6 +77,13 @@ class vmtkImageInitialization(pypes.pypeScript):
             ['InitialLevelSets','olevelsets','vtkImageData',1,'','','vmtkimagewriter'],
             ['Surface','osurface','vtkPolyData',1,'','','vmtksurfacewriter']
             ])
+
+    def _SeedPickCallback(self, caller, event):
+        '''change seed rgb color when two seeds have been selected'''
+        self.NumSeedsSelected = self.ImageSeeder.Seeds.GetNumberOfPoints()
+        if self.NumSeedsSelected % 2 == 0:
+            self.ImageSeeder.SeedRGBColor = [random(), random(), random()]
+            self.ImageSeeder.BuildView()
 
     def SeedInput(self,queryString,numberOfSeeds):
         invalid = 1
@@ -346,6 +355,104 @@ class vmtkImageInitialization(pypes.pypeScript):
 
         self.IsoSurfaceValue = 0.0
 
+    def MultipleCollidingFrontsInitialize(self):
+
+        self.PrintLog('Multiple colliding fronts initialization.')
+        if self.Interactive:
+            queryString = "Please input lower threshold (\'n\' for none): "
+            self.LowerThreshold = self.ThresholdInput(queryString)
+            queryString = "Please input upper threshold (\'n\' for none): "
+            self.UpperThreshold = self.ThresholdInput(queryString)
+
+        scalarRange = self.Image.GetScalarRange()
+        thresholdedImage = self.Image
+
+        if (self.LowerThreshold is not None) | (self.UpperThreshold is not None):
+            threshold = vtk.vtkImageThreshold()
+            threshold.SetInputData(self.Image)
+            if (self.LowerThreshold is not None) & (self.UpperThreshold is not None):
+                threshold.ThresholdBetween(self.LowerThreshold,self.UpperThreshold)
+            elif (self.LowerThreshold is not None):
+                threshold.ThresholdByUpper(self.LowerThreshold)
+            elif (self.UpperThreshold is not None):
+                threshold.ThresholdByLower(self.UpperThreshold)
+            threshold.ReplaceInOff()
+            threshold.ReplaceOutOn()
+            threshold.SetOutValue(scalarRange[0] - scalarRange[1])
+            threshold.Update()
+
+            scalarRange = threshold.GetOutput().GetScalarRange()
[…]
