### 2018-05-19 e735c97
added multiple colliding fronts method to vmtkimageinitialization. added selectable seed color to vmtkimageseeder
