### 2018-05-20 ca0b9d5
Cleaning up code for readability. removing unused methods
