### 2018-05-20 #736 @Kevin-Mattheus-Moerman
@rlizzo @wouterpotters the submitted work is sufficiently advanced with respect to the previous versions to warrant a potential JOSS paper. Future releases of this software, if again significantly enhanced, may also warrant a new additional JOSS paper, since in this way accurate citing and credit giving is enabled.

@wouterpotters thanks for bringing up the point of your contributions to this work in the form of ~2 pull requests. Since these are minor contributions from several years ago, and you are not a co-author or active collaborator currently, I do not see this as a serious source of conflict of interest. It is a benefit that you are technically familiar with this work so will hopefully be able to provided a thorough review. Nevertheless it is good that our review process is open, and multiple reviewers are recruited, such that any issues should come to light.
