### 2018-05-20 b084520
removed legacy pypes testing scripts which are out of use. fixed error where vmtkScripts __init__.py file was being wrapped by pyperun in the Install bin directory
