+++ PypeS/CMakeLists.txt
@@ -4,8 +4,6 @@ set(PYPES_SRCS
   pypebatch.py
   pypepad.py
   pypeserver.py
-  pypetest.py
-  pypetestrunner.py
   pypewrapper.py
   pyperun.py
   )
