### 2018-05-20 1508c2b
removed unused pype method and documentation updates
