### 2018-05-20 702c5a5
extract methods from pypescript in order to declutter. placing all utility functions in a common pypes file
