+++ PypeS/pype.py
@@ -21,30 +21,9 @@ import importlib
 from inspect import isclass
 import re
 import inspect
+from .pypesutils import all_indices
 
 pype = 'Pype'
-
-def all_indices(value, qlist):
-    '''convenience method to a list of all indexes where an element is within a collection
-
-    arguments:
-        value (str, int, float): the item to find within the collection
-        qlist (list, tuple): the collection of elements which you want to find the the indexes
-            of value in.
-
-    returns:
-        indices (list:`int`): a list containing the index of every location where qlist == value.
-    '''
-    indices = []
-    idx = -1
-    while True:
-        try:
-            idx = qlist.index(value, idx+1)
-            indices.append(idx)
-        except ValueError:
-            break
-    return indices
-
 
 class Pype(object):
     '''
@@ -66,13 +45,6 @@ class Pype(object):
         self.InputStream = sys.stdin
         self.OutputStream = sys.stdout
         self.Arguments = None
-
-    def GetUsageString(self):
-        usageString = 'Usage: pype --nolog --noauto firstScriptName -scriptOptionName '\
-                      'scriptOptionValue --pipe secondScriptName -scriptOptionName scriptOptionValue '\
-                      '-scriptOptionName @firstScriptName.scriptOptionName -id 2 --pipe thirdScriptName '\
-                      '-scriptOptionName @secondScriptName-2.scriptOptionName'
-        return usageString
 
     def PrintLog(self,logMessage,indent=0):
         '''Prints log messages from pype controller members to the console.
@@ -151,11 +123,14 @@ class Pype(object):
 
         self.Arguments = arguments
 
-    def _ParseArgumentsFlags(self, arguments):
-        '''execute special case argument flag behavior
+    def _ParsePyperunControlingArgumentFlags(self, arguments):
+        '''execute special case argument flag behavior if the pyperun executable is being prepended to a pype
 
-        check if --help --noauto --nolog --query flags exist in the arguments list. if they do then execute the
-        special behavior indicated by those flags.
+        This was originally used to make parsing more robust. To control the pype behavior or to access global
+        pype functionality like --nolog, the pyperun executable is prepended to a pype.
+
+        checks if --noauto --nolog flags exist in the arguments list. if they do then execute the
+        special behavior indicated by those flags. The --help flag prints the pyperun usage string. 
 
         arguments:
             arguments: (obj:`list`): the initial arguments acquired from sys.argv with each logical unit as a string element.
@@ -171,7 +146,10 @@ class Pype(object):
                 else:
                     break
             if '--help' in pypeArguments:
-                self.PrintLog(self.GetUsageString())
+                self.PrintLog('Usage: pype --nolog --noauto firstScriptName -scriptOptionName '\
+                              'scriptOptionValue --pipe secondScriptName -scriptOptionName scriptOptionValue '\
+                              '-scriptOptionName @firstScriptName.scriptOptionName -id 2 --pipe thirdScriptName '\
+                              '-scriptOptionName @secondScriptName-2.scriptOptionName')
             if '--noauto' in pypeArguments:
                 self.AutoPipe = 0
             if '--nolog' in pypeArguments:
@@ -224,8 +202,8 @@ class Pype(object):
                                ['vmtkimageotsuthresholds', []]]
         '''
         self.ScriptList = []
-        argumentsWithFlagsParsed = self._ParseArgumentsFlags(self.Arguments[:])
-        arguments = self._ParseArgumentsFileBrowser(argumentsWithFlagsParsed)
+        pyperunParsedArguments = self._ParsePyperunControlingArgumentFlags(self.Arguments[:])
+        arguments = self._ParseArgumentsFileBrowser(pyperunParsedArguments)
 
         pipeIndices = all_indices('--pipe', arguments)
[…]
