+++ PypeS/pypescript.py
@@ -21,6 +21,7 @@ import os.path
 import importlib
 from inspect import isclass
 import textwrap
+from .pypesutils import GetUsageString, GetHTMLUsageString
 
 class pypeMember(object):
 
@@ -210,6 +211,8 @@ class pypeScript(object):
         selfMember = pypeMember('Disabled','disabled','bool',1,'','disable execution and piping')
         selfMember.AutoPipe = 0
         self.InputMembers.append(selfMember)
+        self.GetUsageString = GetUsageString
+        self.GetHTMLUsageString = GetHTMLUsageString
 
         self.ExitOnError = 1
         self.LogOn = 1
@@ -450,104 +453,6 @@ class pypeScript(object):
         '''sets the script object ScriptDoc attribute'''
         self.ScriptDoc = scriptDoc
 
-    def GetUsageString(self):
-        '''get a plain text string describing useful information about the current script object
-
-        returns:
-            a formatted string which contains the current ScriptName, ScriptDoc, and Input/Output member
-                fields (name, option, type, length, range, doc, default, and help text)
-        '''
-        usageString = ''
-        scriptUsageString = os.path.splitext(self.ScriptName)[0]
-        if self.ScriptDoc:
-            scriptUsageString += ' : ' + self.ScriptDoc
-        textwrapper = textwrap.TextWrapper()
-        textwrapper.initial_indent = ''
-        textwrapper.subsequent_indent = ' '
-        usageString += textwrapper.fill(scriptUsageString)
-        for memberList in [self.InputMembers, self.OutputMembers]:
-            if memberList == self.InputMembers:
-                 usageString += '\n' + '  ' + 'Input arguments:'
-            elif memberList == self.OutputMembers:
-                 usageString += '\n' + '  ' + 'Output arguments:'
-            for memberEntry in memberList:
-                memberUsageString = ''
-                indentation = '   '
-                textwrapper.initial_indent = indentation
-                textwrapper.subsequent_indent = indentation + '  '    
-                memberName  = memberEntry.MemberName
-                option = memberEntry.OptionName
-                memberType = memberEntry.MemberType
-                memberLength = memberEntry.MemberLength
-                memberRange = memberEntry.MemberRange
-                memberDoc = memberEntry.MemberDoc
-                if option == '':
-                    usageString += '\n' + textwrapper.fill(memberUsageString)
-                    continue
-                if memberLength == 0:
-                    memberUsageString += '-' + option + ' ' + memberName
-                else:
-                    memberUsageString += '-' + option + ' ' + memberName + ' (' + memberType + ',' + str(memberLength) + ')'
-                    if (memberRange) and (memberType in self.BuiltinOptionTypes):
-                        memberUsageString += " " + memberEntry.GetRangeRepresentation()
-                    if (getattr(self, memberName) != '') and (memberType in self.BuiltinOptionTypes):
-                        memberUsageString += '; default=' + str(getattr(self, memberName))
-                if memberDoc != '':
-                    memberUsageString += ': ' + memberDoc
-                usageString += '\n' + textwrapper.fill(memberUsageString)
-        usageString += '\n'
-        return usageString
-
-    def GetHTMLUsageString(self):
-        '''get a html formatted tagged string describing useful information about the current script object
-
-        returns:
-            A formatted string which contains the current ScriptName, ScriptDoc, and Input/Output member fields
-                (name, option, type, length, range, doc, default, and help text)
-        '''
-        usageString = ''
-        usageString += '<h1>'
-        usageString += self.ScriptName
-        usageString += '</h1>'
-        usageString += '\n'
-        if self.ScriptDoc != '':
-            usageString += '<h2>Description</h2>' + '\n'
[…]
