### 2018-05-21 #745 @whedon
[REVIEW]: The Vascular Modeling Toolkit: A Python Library for the Analysis of Tubular Structures in Medical Images
**Submitting author:** @rlizzo (<a href="http://orcid.org/0000-0002-0811-6513">Richard Izzo</a>)
**Repository:** <a href="https://github.com/vmtk/vmtk" target ="_blank">https://github.com/vmtk/vmtk</a>
**Version:** v1.4.0
