### 2018-05-21 #745 @Kevin-Mattheus-Moerman
@rlizzo, @wouterpotters, @brainstorm this is where the review process takes place. There are sets of tickboxes for both reviewers at the top of this issue (let me know if you are unable to tick them). If you have questions please let me know. You may also want to check out the [JOSS reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines).
