### 2018-05-21 #745 @Kevin-Mattheus-Moerman
@osmsc @sshadden @jmerkow @alexkaiser @updega2 @gmaher @fkong7 this is a [JOSS](http://joss.theoj.org/about) review issue for [ _The Vascular Modeling Toolkit_ ](https://github.com/vmtk/vmtk). Please let me know if you are interested in reviewing this submission.

The review process is focused largely on the software and a short paper, see the [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines).

[--> Check software repository :floppy_disk: <--](https://github.com/vmtk/vmtk)
