### 2018-05-21 vmtk/vmtk#283 @wouterpotters
JOSS - License in LICENSE file
According to the [JOSS reviewer guidelines](https://joss.theoj.org/about#reviewer_guidelines), the VMTK licence should be placed in a plain-text LICENSE file. Currently the license is in the copyright.txt file.
