### 2018-05-21 #745 @wouterpotters
minor issue: license in LICENSE file https://github.com/vmtk/vmtk/issues/283
