### 2018-05-21 vmtk/vmtk#283 @rlizzo
Thanks for the heads up!

@lantiga, is there any reason the license/copyright needs to be structured this way? Or are we good to move all the text into the LICENSE file?
