### 2018-05-21 vmtk/vmtk#283 @rlizzo
Thanks. Also @wouterpotters there is a separate (open source compatible) LICENCE file for the PypeS module https://github.com/vmtk/vmtk/blob/master/PypeS/LICENCE . Does this need to be included in to top level LICENSE as well?
