### 2018-05-21 vmtk/vmtk#283 @wouterpotters
just a small addition to my previous remark: http://www.vmtk.org/license.html and the GitHub LICENSE seem to be dated at 2011 and/or 2012. It would be neat to update all the dates to 'until 2018' as well I guess.
