### 2018-05-21 vmtk/vmtk#283 @wouterpotters
To be honest I'm no expert on the licenses, but it would be convenient for end users to see all the relevant licenses for the 1 open source software package in 1 place (the LICENSE file).
