### 2018-05-21 vmtk/vmtk#283 @lassoan
A single LICENSE file in a standard location is important because applications that use many libraries often harvest license files from library repositories automatically and append them to the application's license file that is displayed to users. If you use some custom filename or location then application developers need to manually create a rule for finding license information for your library, or license information for your library may be just skipped.

Also, your library may not get listed in online software repositories or score low if license file cannot be found automatically.
