### 2018-05-21 4bd95e2
[ci skip] moved all license and copyright info into a top level LICENSE file, updated library files to reflect change in name and location
