+++ vmtkScripts/vmtkcenterlinelabeler.py
@@ -7,7 +7,7 @@
 ## Version:   $Revision: 1.2 $
 
 ##   Copyright (c) Luca Antiga, David Steinman. All rights reserved.
-##   See LICENCE file for details.
+##   See LICENSE file for details.
 
 ##      This software is distributed WITHOUT ANY WARRANTY; without even 
 ##      the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR 
