+++ vtkVmtk/Segmentation/vtkvmtkCurvesLevelSetImageFilter.h
@@ -7,7 +7,7 @@ Date:      $Date: 2006/04/06 16:48:25 $
 Version:   $Revision: 1.4 $
 
   Copyright (c) Luca Antiga, David Steinman. All rights reserved.
-  See LICENCE file for details.
+  See LICENSE file for details.
 
   Portions of this code are covered under the VTK copyright.
   See VTKCopyright.txt or http://www.kitware.com/VTKCopyright.htm 
