+++ vtkVmtk/ComputationalGeometry/vtkvmtkPolyDataFlowExtensionsFilter.h
@@ -7,7 +7,7 @@ Date:      $Date: 2006/07/07 10:46:19 $
 Version:   $Revision: 1.6 $
 
   Copyright (c) Luca Antiga, David Steinman. All rights reserved.
-  See LICENCE file for details.
+  See LICENSE file for details.
 
   Portions of this code are covered under the VTK copyright.
   See VTKCopyright.txt or http://www.kitware.com/VTKCopyright.htm 
