+++ distribution/travis/install.sh
@@ -4,7 +4,7 @@
 ## Version:   1.4
 ##
 ##   Copyright (c) Richard Izzo, Luca Antiga, All rights reserved.
-##   See LICENCE file for details.
+##   See LICENSE file for details.
 ##
 ##      This software is distributed WITHOUT ANY WARRANTY; without even
 ##      the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
