### 2018-05-21 vmtk/vmtk#283 @rlizzo
Thanks for the feedback @wouterpotters and @lassoan! I've made the requested modifications to the repository and vmtk website and have linked the PRs to this issue and the main review issue.

@wouterpotters, I'll leave this to you to close once you confirm that this is now in compliance.
