### 2018-05-21 vmtk/vmtk#283 @wouterpotters
Looks good now 👍 closing this one.
