### 2018-05-21 vmtk/vmtk#286 @wouterpotters
JOSS - Installation instructions 
For non-anaconda Mac users, please add this line (or something similar) to the installation instructions: "Start the newly installed vmtk anaconda environment using `source activate foo`"
