### 2018-05-21 vmtk/vmtk#286 @wouterpotters
also, please include an example where the use of double quotes in pypepad (''/path/to/dicomfiles'' in stead of '/path/to/dicomfiles') is illustrated. This avoids a lot of user confusion for new vmtk users.
