### 2018-05-21 vmtk/vmtk#286 @rlizzo
Hi @wouterpotters, I just addressed the install instructions.

For your second point can you link to the tutorial you're referring to (where  '/path/to/dicomfiles' is illustrated)?
