### 2018-05-21 vmtk/vmtk#286 @wouterpotters
I meant this one. I remember several first time users not understanding the fact that two '' are required to encompass a path with spaces (would be good to mention this in the 'getting started' page in order to not deter new users: http://www.vmtk.org/documentation/getting-started.html)

PS: please note that no ' are used on that page right now, but this example only works when a user has no spaces.
