### 2018-05-21 vmtk/vmtk#286 @rlizzo
Thanks! Addressed in https://github.com/vmtk/vmtk.github.com/pull/14
