### 2018-05-21 vmtk/vmtk#288 @wouterpotters
JOSS - Incomplete documentation
The documentation for vmtk could be improved:

1. Create an description for every python script (https://www.vmtk.org/documentation/vmtkscripts.html is almost complete)

2. http://www.vmtk.org/doc/html/ for the C++ classes is incomplete. There are very few items that have user friendly help information here (besides the automatically generated content). It should be rather straightforward to add more documentation (i.e. human readable explanations for the classes) to the C code, so that contributors quickly have an idea what the file is supposed to do; doxygen has some features to automate most of this: http://www.stack.nl/~dimitri/doxygen/manual/docblocks.html, although manual typing is still required (once).
