### 2018-05-21 #745 @wouterpotters
There are no automated tests for the vmtk Python scripts and C++ classes.
However:

- A large part of the software is based on ITK routines, which do have automated testing for all their routines.
- There are options for visual user feedback (i.e. ITK visualisations) every step of the way using vmktimageviewer, vmtksurfaceviewer, vmtkrenderer
- There are options for checking the results in other software packages (e.g. by using vmtksurfacewriter and vmtkimagewriter).
- the website does have several extensive tutorials and example data using the visual feedback.

I would consider the added value for automated vmtk tests limited. I would advise to look into this, but I do not consider it a requirement for publication. What do you think @brainstorm?
