### 2018-05-21 vmtk/vmtk#288 @rlizzo
Ahh.. updating the doxygen build and C++ code documentation has been on my to-do list for way to long. Glad to finally have a reason to spend time on it.

@lantiga, let's discuss the best way to do this.
