### 2018-05-21 d752b50
work on increasing test coverage of pypes module
