+++ tests/test_PypeS/test_pype.py
@@ -16,11 +16,40 @@
 
 import pytest
 import sys
+import vtk
 from vmtk import pypes
 
 def test_init():
     pipe = pypes.Pype()
     assert pipe.ScriptName == 'pype'
+
+#---------------------------------------------------
+# Test PrintLog Method
+#---------------------------------------------------
+def test_print_log_no_indent(capsys):
+    pipe = pypes.Pype()
+    pipe.PrintLog('testing no indent')
+    captured = capsys.readouterr()
+    assert captured.out == 'testing no indent\n'
+
+
+def test_print_log_one_indent(capsys):
+    pipe = pypes.Pype()
+    pipe.PrintLog('testing one indent', 1)
+    captured = capsys.readouterr()
+    assert captured.out == '    testing one indent\n'
+
+
+def test_print_log_set_logon_off(capsys):
+    pipe = pypes.Pype()
+    pipe.LogOn = 0
+    pipe.PrintLog('this should not be printed')
+    captured = capsys.readouterr()
+    assert captured.out == ''
+
+#---------------------------------------------------
+# Test SetArgumentString Method
+#---------------------------------------------------
 
 def test_set_argument_string_two_args():
     pipe = pypes.Pype()
@@ -36,13 +65,87 @@ def test_set_argument_string_five_args():
     assert pipe.Arguments == ['vmtkimagereader', '--help', '0123', 'vmtkimageviewer', 'fixit']
 
 
-def test_fail_set_argument_string_nonmatching_quote():
+def test_set_argument_string_matching_double_quote_no_space_in_file_name():
+    pipe = pypes.Pype()
+    arg_string = 'vmtkimagereader -ifile "/foo/anotherdirectory/bar.vti" --pipe vmtkimageviewer'
+    pipe.SetArgumentsString(arg_string)
+    assert pipe.Arguments == ['vmtkimagereader', '-ifile', '/foo/anotherdirectory/bar.vti', '--pipe', 'vmtkimageviewer']
+
+
+def test_set_argument_string_matching_double_quote_space_in_file_name():
+    pipe = pypes.Pype()
+    arg_string = 'vmtkimagereader -ifile "/foo/another directory/bar.vti" --pipe vmtkimageviewer'
+    pipe.SetArgumentsString(arg_string)
+    assert pipe.Arguments == ['vmtkimagereader', '-ifile', '/foo/another directory/bar.vti', '--pipe', 'vmtkimageviewer']
+
+
+def test_fail_set_argument_string_nonmatching_double_quote(capsys):
     with pytest.raises(SystemExit) as pytest_wrapped_e:
         pipe = pypes.Pype()
         arg_string = 'vmtkimagereader --help"'
         pipe.SetArgumentsString(arg_string)
     assert pytest_wrapped_e.type == SystemExit
+    captured = capsys.readouterr()
+    # need newline on comparison since pype.PrintError automatically includes it. 
+    assert captured.out == 'Error: non-matching quote found\n'
 
+
+def test_set_argument_string_no_quote_space_in_file_name():
+    # this is not the behavior we want long term, since it split up the file name, but the set argument string
+    # method is not responsible for catching that
+    pipe = pypes.Pype()
+    arg_string = 'vmtkimagereader -ifile /foo/another directory/bar.vti --pipe vmtkimageviewer'
+    pipe.SetArgumentsString(arg_string)
[…]
