### 2018-05-21 vmtk/vmtk#286 @wouterpotters
Fixed
