### 2018-05-21 #745 @wouterpotters
Extensive testing; all explained in the CONTRIBUTING.md file, test folder and on https://github.com/vmtk/vmtk-test-data.
