### 2018-05-21 #745 @wouterpotters
already fixed; new user instructions: https://github.com/vmtk/vmtk/issues/286
