### 2018-05-21 #745 @wouterpotters
Despite the fact that there already is a lot of useful documentation for both users and developers: issue with (incomplete) documentation: http://github.com/vmtk/vmtk/issues/288

Improving this point would improve the open source aspect of vmtk
