### 2018-05-21 vmtk/vmtk#289 @wouterpotters
JOSS - paper.md
Some comments and suggestions for the accompanying paper.md:
> "python-wrapped C++ methods"
--> change into C++ classes?
> "Image segmentation is performed via"
--> change into "can be performed"? (thresholding also is an option in VMTK)
> "unique system of "
--> change into "unique interface of"?

For all figures in the paper.md: please explain what you see in the figures, label the subfigures if required. This information is currently in the text rather than in the caption. (which is confusing as the figures are moved around).

Finally the first and the 3rd author (@rizzo and @daron1337) are not mentioned in all the README, LICENSE and other files, but they are on the paper. I think it would make sense to add them to all the notices in the repository as well (also considering their substantial contributions to the vmtk repository).
