### 2018-05-21 #745 @wouterpotters
Some small issues with the submitted manuscript: https://github.com/vmtk/vmtk/issues/289
