### 2018-05-21 vmtk/vmtk#289 @wouterpotters
The affiliations include titles (e.g. professor); this is not common. Please only refer to the institutions and use prefixes for titles.
