### 2018-05-21 #745 @wouterpotters
Overall opinion on the submitted paper:
The VMTK software package is a scientifically relevant open source software package that is relevant specifically to the field of vascular imaging and simulation.

The package is useful for users at different levels. Clinicians and non-technical users can use their precompiled version, more advanced developers can use either the python or the C++ libraries.

As the package is largely based on the common VTK and ITK packages, the software is easily used within other VTK / ITK based packages and the other way around. VTK / ITK features are easily implemented in vmtk.

Key features besides VTK / ITK are the pypes pipeline, which enables users to manually or automatically process all types of files into custom pipelines with custom output files.
Another key feature is the unique and robust way to detect and process the centrelines of tubular structures. Together with several surface processing and mesh generating functions, this software package can quantify basic vascular/tubular properties and perform the required preparations for (open source) CFD simulations.
The extensive manuals and (almost complete) documentation pages guide new user through these features.

The previous release of this software (2008) has been cited and used a lot by many users. In the past years the software package went through some significant and important changes (test system, user interface improvements, new features).

Overall I would advise to accept the publication of this excellent open source software package, provided that the authors improve the C++ classes documentation and the paper.md.
