### 2018-05-21 vmtk/vmtk#289 @rlizzo
Thanks for the feedback. I'll make the modifications tonight or tomorrow.

Question for @wouterpotters or @Kevin-Mattheus-Moerman, is there a way for me to generate the PDF locally so I can preview the output and ensure that figures/captions are placed correctly? I don't want to clutter the review thread with calls to "Wheaton generate PDF" if I have to go through a few iterations to get the document structure tidied up.
