### 2018-05-22 vmtk/vmtk#289 @Kevin-Mattheus-Moerman
@rlizzo I think you can try:
`pandoc --filter pandoc-citeproc --bibliography paper.bib  paper.md -o paper.pdf`

See also here: https://github.com/openjournals/whedon/issues/17
