### 2018-05-22 a0d037a
joss review - updated paper with review comments, included doi in bibliography, get a better figure for pypepad, fixed figure position on page, updated author affiliation, and include myself (Richard Izzo) and Simone Manini as copyright holders in the LICENSE file as per joss reviewer comments
