### 2018-05-22 vmtk/vmtk#289 @rlizzo
Thanks @Kevin-Mattheus-Moerman, that worked well for me! Glad it did, since it took a few tries to get it all right.

@wouterpotters, I've updated the paper to reflect your changes. please let me know if this is sufficient.
