### 2018-05-22 9c1bbe8
joss - added missing doc strings for vmtkscripts
