+++ vmtkScripts/vmtklevelsetsegmentation.py
@@ -72,6 +72,7 @@ class vmtkLevelSetSegmentation(pypes.pypeScript):
         self.SmoothingConductance = 0.8
 
         self.SetScriptName('vmtklevelsetsegmentation')
+        self.SetScriptDoc('interactivly initialize an initial level set and evolve it to image gradients')
         self.SetInputMembers([
             ['Image','i','vtkImageData',1,'','','vmtkimagereader'],
             ['FeatureImage','featureimage','vtkImageData',1,'','','vmtkimagereader'],
