+++ vmtkScripts/vmtkflowextensions.py
@@ -50,6 +50,7 @@ class vmtkFlowExtensions(pypes.pypeScript):
         self.OwnRenderer = 0
 
         self.SetScriptName('vmtkflowextensions')
+        self.SetScriptDoc('interactivly add flow extensions to a surface for CFD processing')
         self.SetInputMembers([
             ['Surface','i','vtkPolyData',1,'','','vmtksurfacereader'],
             ['Centerlines','centerlines','vtkPolyData',1,'','','vmtksurfacereader'],
