### 2018-05-22 vmtk/vmtk#288 @rlizzo
Hi @wouterpotters, just making you aware I've addressed point 1 of https://github.com/vmtk/vmtk/issues/288#issue-325029959
