### 2018-05-22 9fe869e
DOC, WIP: Initial work adding C++ docstrings.
This is requested by the JOSS reviewers for paper acceptance
