+++ vtkVmtk/Segmentation/vtkvmtkMedialCurveFilter.h
@@ -26,9 +26,18 @@ Version:   $Revision: 1.4 $
 
 =========================================================================*/
 
-// .NAME vtkvmtkMedialCurveFilter.h - Wrapper class around
+// .NAME vtkvmtkMedialCurveFilter - flux driven automatic centerline extraction from a surface
 // .SECTION Description
-// vtkvmtkMedialCurveFilter.h
+// Please see http://www.insight-journal.org/browse/publication/165 for source implementation. 
+// Implementation of the flux driven automatic centerline extraction
+// algorithm proposed by Bouix et al. in 2004. This is based on a skeletonisation algorithm initially
+// proposed by Siddiqi et al. in 2002, using properties of an average outward flux measure to distinguish
+// skeletal points from non-skeletal ones. This information is combined with a topology preserving thinning
+// procedure to obtain the final result. This implementation combines this skeletonisation algorithm with
+// other techniques as described in the paper of Bouix et al. to produce an ITK filter that generates as output
+// the skeleton, as a binary object represented in an image, of the input surface, represented as a distance
+// transform image. In this work is described this medial curve extraction procedure following the ITK
+// philosophy.
 
 
 #ifndef __vtkvmtkMedialCurveFilter_h
