+++ vtkVmtk/Misc/vtkvmtkSurfaceDistance.h
@@ -18,7 +18,7 @@ Version:   $Revision: 1.4 $
      PURPOSE.  See the above copyright notices for more information.
 
 =========================================================================*/
-  // .NAME vtkvmtkSurfaceDistance - ...
+  // .NAME vtkvmtkSurfaceDistance - calculate the euclidian distance of one surface from another
   // .SECTION Description
   // .
 
