+++ vtkVmtk/Misc/vtkvmtkPolyDataToUnstructuredGridFilter.h
@@ -18,7 +18,7 @@ Version:   $Revision: 1.4 $
      PURPOSE.  See the above copyright notices for more information.
 
 =========================================================================*/
-  // .NAME vtkvmtkPolyDataToUnstructuredGridFilter - ...
+  // .NAME vtkvmtkPolyDataToUnstructuredGridFilter - create a mesh type object from a surface type object
   // .SECTION Description
   // ...
 
