### 2018-05-23 ed0547f
added cmake gui option to build vmtk doxygen documentation. Added documentation generation instructions to the contributor guidelines
