+++ CONTRIBUTING.md
@@ -111,7 +111,18 @@ We use the [pytest](https://docs.pytest.org/en/latest/) testing framework for un
 
 In order to contribute tests, can clone the [vmtk-test-data](https://github.com/vmtk/vmtk-test-data) repository at the same level as vmtk. Please add any test data files in the appropriate folders in the `vmtk-test-data` repository. if you create new test files within the `tests/test_FOO` directory, be sure to add that file to the accompanying `CMakeLists.txt` file as well. When your tests run locally, your data assets and test changes should be reflected and all tests should appear to pass. 
 
-If you are contributing tests which require new data files, create a PR to the [vmtk-test-data](https://github.com/vmtk/vmtk-test-data) repository and let us know the PR# in VMTK which the data files corresponds to. Until we merge the data, our CI suite will not have access to files, so your tests that passed locally may appear to fail when they are pushed - Don't worry! It's not you, it's our system! 
+If you are contributing tests which require new data files, create a PR to the [vmtk-test-data](https://github.com/vmtk/vmtk-test-data) repository and let us know the PR# in VMTK which the data files corresponds to. Until we merge the data, our CI suite will not have access to files, so your tests that passed locally may appear to fail when they are pushed - Don't worry! It's not you, it's our system!
+
+## Building Documentation
+
+### C++ Class Documentation
+
+The [VMTK C++ class documentation](http://www.vmtk.org/doc/html/index.html) is generated by [Doxygen](http://www.stack.nl/~dimitri/doxygen/index.html). The system requires that you have Doxygen, GraphViz, and a LaTeX complier set up on your system path. To enable the html documentation build, set the CMake variable `-DBUILD_DOCUMENTATION=ON` from the command line (or toggle `BUILD_VMTK_DOCUMENTATION` to `ON` if using the CMake curses interface or CMake GUI). The fully configured `doxyfile` and `doc_makeall.sh` file will be placed in the build-directory under the path `foo/VMTK-Build/vtkVmtk/Utilities/Doxygen/` (by default). To generate the website, run the `foo/VMTK-Build/vtkVmtk/Utilities/Doxygen/doc_makeall.sh` file from the terminal. The html files and other resources are placed in `foo/VMTK-Build/vtkVmtk/Utilities/Doxygen/doc/html/` directory. To update the [vmtk.org](www.vmtk.org) website documentation, clone the [vmtk.github.com](https://github.com/vmtk/vmtk.github.com) repository, copy the files into the corresponding `doc/html` directory, commit the files, and propose the changes as a pull request.
+
+### Python vmtkScripts Documentation
+
+[Python vmtkScripts documentation](http://www.vmtk.org/documentation/vmtkscripts.html) is generated by [the vmtkscripts2html.sh shell script](https://github.com/vmtk/vmtk.github.com/blob/master/vmtkscripts2html.sh) contained at the root level of the [vmtk.github.com](https://github.com/vmtk/vmtk.github.com) repository. To use this utility, you first need to build (from source) the version of the VMTK library you want to document. Activate that build in your current terminal session, and `cd` into the top level of your cloned `vmtk.github.com` folder. Call the `vmtkscripts2html.sh` with a single argument: the path to `Install/bin` directory of the local copy of VMTK you built (ex: `bash ./vmtkscripts2html.sh ~/projects/vmtk-build/Install/bin`). This may take a few minutes to run, but it will automatically generate the index/summary page and a corresponding detailed description page for each vmtkscript in the library. It will additionally place them in the appropriate places within the vmtk.github.com folder. Simply commit these changes and propose them as a pull request to update the website documentation.
 
 ## Questions? Concerns?
[…]
