+++ vtkVmtk/Utilities/Doxygen/doxyfile.in
@@ -49,9 +49,9 @@ EXTRACT_ALL          = YES
 EXTRACT_PRIVATE      = NO
 EXTRACT_STATIC       = YES
 HIDE_UNDOC_MEMBERS   = NO
-HIDE_UNDOC_CLASSES   = YES
+HIDE_UNDOC_CLASSES   = NO
 ALWAYS_DETAILED_SEC  = NO
-SOURCE_BROWSER       = NO
+SOURCE_BROWSER       = YES
 INLINE_SOURCES       = NO
 CASE_SENSE_NAMES     = YES
 VERBATIM_HEADERS     = NO
