### 2018-05-23 6ffd2d1
updated documentation and doxygen file config
