+++ vtkVmtk/IO/vtkvmtkFluentWriter.h
@@ -18,11 +18,10 @@ Version:   $Revision: 1.2 $
      PURPOSE.  See the above copyright notices for more information.
                                                                                                                                     
 =========================================================================*/
-
 // .NAME vtkvmtkFluentWriter - writes Fluent .msh files.
 // .SECTION Description
 // vtkvmtkFluentWriter writes Fluent .msh files. Many thanks to M. Xenos, Y. Alemu and D. Bluestein, BioFluids Laboratory, Stony Brook University, Stony Brook, NY, for the inputs on the file format.
-
+//
 // .SECTION See Also
 
 #ifndef __vtkvmtkFluentWriter_h
