+++ vtkVmtk/ComputationalGeometry/vtkvmtkAppendFilter.h
@@ -18,9 +18,9 @@ Version:   $Revision: 1.4 $
      PURPOSE.  See the above copyright notices for more information.
 
 =========================================================================*/
-  // .NAME vtkvmtkAppendFilter - merges points, point data, and cell data
-  // .SECTION Description
-  // .
+// .NAME vtkvmtkAppendFilter - merges points, point data, and cell data.
+// .SECTION Description
+// ...
 
 // The RequestData method implementation is heavily based on the same method in vtkAppendFilter,
 // which is covered by the following copyright notice.
