+++ vtkVmtk/IO/vtkvmtkXdaReader.h
@@ -18,11 +18,10 @@ Version:   $Revision: 1.3 $
      PURPOSE.  See the above copyright notices for more information.
                                                                                                                                     
 =========================================================================*/
-
 // .NAME vtkvmtkXdaReader - reads libmesh Xda files
 // .SECTION Description
 // vtkvmtkXdaReader reads libmesh Xda files
-
+//
 // .SECTION See Also
 
 #ifndef __vtkvmtkXdaReader_h
