### 2018-05-24 9ce7607
finished documentation updates to vtkvmtkComputationalGeometry.
fixed comment style for doxygen generation
