+++ vtkVmtk/ComputationalGeometry/vtkvmtkCapPolyData.h
@@ -18,7 +18,7 @@ Version:   $Revision: 1.4 $
      PURPOSE.  See the above copyright notices for more information.
 
 =========================================================================*/
-// .NAME vtkvmtkCapPolyData - Add caps to boundaries
+// .NAME vtkvmtkCapPolyData - Close holes in a surface surface by creating a cap made of triangles sharing a common vertex at the boundary barycenter.
 // .SECTION Description
 // This class closes the boundaries of a surface with a cap. Each cap is made of triangles sharing the boundary baricenter. Boundary baricenters are added to the dataset. It is possible to retrieve the ids of the added points with GetCapCenterIds. Boundary baricenters can be displaced along boundary normals through the Displacement parameter. Since this class is used as a preprocessing step for Delaunay tessellation, displacement is meant to avoid the occurence of degenerate tetrahedra on the caps.
 
