+++ vtkVmtk/ComputationalGeometry/vtkvmtkPolyDataCenterlineAbscissaMetricFilter.h
@@ -18,9 +18,9 @@
      PURPOSE.  See the above copyright notices for more information.
 
 =========================================================================*/
-// .NAME vtkvmtkPolyDataCenterlineAbscissaMetricFilter - ..
+// .NAME vtkvmtkPolyDataCenterlineAbscissaMetricFilter - Project the abscissa metric of the nearest centerline point to every point of a surface.  
 // .SECTION Description
-// ..
+// Briefly, each segment of a vascular network is topologically equivalent to a cylinder and can consequently be mapped onto a rectangular parametric space that allows both easier investigations and comparisons between different models and datasets. The parameterization is performed longitudinally by means of the curvilinear abscissa computed over the model centerlines and circumferentially, by the angular position of each point on the surface mesh with respect to the centerlines.
 
 #ifndef __vtkvmtkPolyDataCenterlineAbscissaMetricFilter_h
 #define __vtkvmtkPolyDataCenterlineAbscissaMetricFilter_h
