+++ vtkVmtk/Misc/vtkvmtkAnnularCapPolyData.h
@@ -20,8 +20,7 @@ Version:   $Revision: 1.5 $
 =========================================================================*/
 // .NAME vtkvmtkAnnularCapPolyData - Add annular caps between the boundaries of a walled surface.
 // .SECTION Description
-// This class closes the boundaries between the surfaces of a walled surface with caps. The
-// surfaces are required to be dense for the algorithm to produce legal caps.
+// This class closes the boundaries between the surfaces of a walled surface with caps. The surfaces are required to be dense for the algorithm to produce legal caps.
 
 #ifndef __vtkvmtkAnnularCapPolyData_h
 #define __vtkvmtkAnnularCapPolyData_h
