### 2018-05-24 #745 @wouterpotters
Already looking good now with the latest additions. As soon as the C++ docs are (more) complete, I vote to accept the package. What do you think @brainstorm?
