+++ vtkVmtk/Utilities/Doxygen/doxyfile.in
@@ -28,7 +28,7 @@ INCLUDED_BY_GRAPH    = YES
 CLASS_DIAGRAMS       = YES
 GENERATE_LEGEND      = YES
 GRAPHICAL_HIERARCHY  = YES
-MAX_DOT_GRAPH_DEPTH  = 3
+MAX_DOT_GRAPH_DEPTH  = 5
 
 ALLEXTERNALS         = NO
 
@@ -62,7 +62,6 @@ JAVADOC_AUTOBRIEF    = YES
 SORT_MEMBER_DOCS     = NO
 DISTRIBUTE_GROUP_DOC = YES
 TAB_SIZE             = 3
-# PAPER_TYPE           = letter
 CLANG_ASSISTED_PARSING = YES
 
 FILE_PATTERNS        = *.h *.md
@@ -70,7 +69,7 @@ RECURSIVE            = YES
 
 ALPHABETICAL_INDEX   = YES
 COLS_IN_ALPHA_INDEX  = 3
-IGNORE_PREFIX        = vtk
+IGNORE_PREFIX        = vtkvmtk
 
 ENABLE_PREPROCESSING = YES
 MACRO_EXPANSION      = YES
