### 2018-05-24 efff1a2
basic documentation of the Differential geometry package
