+++ vtkVmtk/DifferentialGeometry/vtkvmtkDataSetItems.h
@@ -18,7 +18,7 @@
      PURPOSE.  See the above copyright notices for more information.
 
 =========================================================================*/
-// .NAME vtkvmtkDataSetItems - ..
+// .NAME vtkvmtkDataSetItems - Create a collection of datat set items, one for every point within the data set.
 // .SECTION Description
 // ..
 
