### 2018-05-24 39ab42d
typo fixes
