+++ vtkVmtk/ComputationalGeometry/vtkvmtkPolyBallLine.h
@@ -22,7 +22,7 @@ Version:   $Revision: 1.4 $
 // .SECTION Description
 //  See detailed description of implicit function in the documentation for vtkvmtkPolyBall class. 
 //
-//  Similar to vtkvmtkPolyBall, the core function of this class is to evaluate the minimum sphere function from on input centerline with associated shere radii and a query point location. Unlike vtkvmtkPolyBall, this class constructs a continuous tubular envelope whose shape is defined by the linear linear interpolation of the circular boundary profiles (with radius equal to the associated shere radii) between every consecutive point on the line. As the boundary profiles are constructed from the centerline sphere radii, the tubular envelope generated is guarrenteed to lie completely within the surface volume. When evaluated, this is essentially equivalent to evaluating a polyball function for an infinitly large collection of spheres along an input dataset. 
+//  Similar to vtkvmtkPolyBall, the core function of this class is to evaluate the minimum sphere function from on input centerline with associated sphere radii and a query point location. Unlike vtkvmtkPolyBall, this class constructs a continuous tubular envelope whose shape is defined by the linear linear interpolation of the circular boundary profiles (with radius equal to the associated sphere radii) between every consecutive point on the line. As the boundary profiles are constructed from the centerline sphere radii, the tubular envelope generated is guaranteed to lie completely within the surface volume. When evaluated, this is essentially equivalent to evaluating a polyball function for an infinity large collection of spheres along an input dataset. 
 
 #ifndef __vtkvmtkPolyBallLine_h
 #define __vtkvmtkPolyBallLine_h
