+++ vtkVmtk/ComputationalGeometry/vtkvmtkPolyDataCenterlineSections.h
@@ -20,12 +20,12 @@ Version:   $Revision: 1.1 $
 =========================================================================*/
 // .NAME vtkvmtkPolyDataCenterlineSections - Extract a set of vessel sections from a surface along a centerline.
 // .SECTION Description
-// The set of vessel sections contain the profile as well as the the following information about the section:
-//     - Centerline Section Area
-//     - Centerline Section Min Size
-//     - Centerline Section Max Size
-//     - Centerline Section Shape
-//     - Centerline Section Closed
+//  The set of vessel sections contain the profile as well as the the following information about the section:
+//  - Centerline Section Area
+//  - Centerline Section Min Size
+//  - Centerline Section Max Size
+//  - Centerline Section Shape
+//  - Centerline Section Closed
 
 #ifndef __vtkvmtkPolyDataCenterlineSections_h
 #define __vtkvmtkPolyDataCenterlineSections_h
