### 2018-05-24 3011c4f
add obj writer in preparation for upgraded vtk
