### 2018-05-25 #745 @brainstorm
Yeah, likewise @wouterpotters, the functional docs are the only area where a bit more text/work would be great to have to pass this review, imho.
