### 2018-05-25 304c868
[ci skip] Use this commit to generate C++ Class VMTK Webpage
Update some of the finite element solver assembler documentation, some
segmentation documentation, and fixup a few typos. At this point most of
the C++ class reference (that we want people playing with) should have
sufficient explainations attached to them.
