+++ vtkVmtk/DifferentialGeometry/vtkvmtkPolyDataFELaplaceAssembler.h
@@ -18,7 +18,7 @@
      PURPOSE.  See the above copyright notices for more information.
 
 =========================================================================*/
-// .NAME vtkvmtkPolyDataFELaplaceAssembler - ..
+// .NAME vtkvmtkPolyDataFELaplaceAssembler - construct a laplacian based finite element calculation.
 // .SECTION Description
 // ..
 
