+++ vtkVmtk/ComputationalGeometry/vtkvmtkAppendFilter.h
@@ -21,8 +21,7 @@ Version:   $Revision: 1.4 $
 // .NAME vtkvmtkAppendFilter - merges points, point data, and cell data.
 // .SECTION Description
 // The RequestData method implementation is heavily based on the same method in vtkAppendFilter.
-
-  // vtkAppendFilter is covered by the following copyright notice.
+// vtkAppendFilter is covered by the following (VTK) copyright notice.
 /*=========================================================================
 
   Program:   Visualization Toolkit
