+++ vtkVmtk/Misc/vtkvmtkSurfMeshWrapper.h
@@ -18,7 +18,7 @@ Version:   $Revision: 1.4 $
      PURPOSE.  See the above copyright notices for more information.
 
 =========================================================================*/
-// .NAME vtkvmtkSurfMeshWrapper - Converts linear elements to quadratic.
+// .NAME vtkvmtkSurfMeshWrapper - ..
 // .SECTION Description
 // ...
 
