+++ vtkVmtk/Utilities/Doxygen/doxyfile.in
@@ -18,7 +18,7 @@ HAVE_DOT             = YES
 DOT_PATH             = @DOXYGEN_DOT_PATH@
 DOT_IMAGE_FORMAT     = svg
 INTERACTIVE_SVG      = YES
-HTML_FILE_EXTENSION  = .xhtml
+HTML_FILE_EXTENSION  = .html
 CLASS_GRAPH          = YES
 COLLABORATION_GRAPH  = YES
 TEMPLATE_RELATIONS   = YES
