### 2018-05-25 #745 @rlizzo
@brainstorm and @wouterpotters: Just letting you know that we have merged the improvements to the C++ class level documentation you requested. You can review the the changes in the Merged PR https://github.com/vmtk/vmtk/pull/295, or view the additions online: http://www.vmtk.org/doc/html/annotated.html. The majority of effort went into documenting the code which has to do with data structures or processing methods which are more specialized to the field of computational vascular analysis. For your reference, the centerline analysis scripts primarily exist within the [vtkVmtk/ComputationalGeometry](https://github.com/vmtk/vmtk/tree/master/vtkVmtk/ComputationalGeometry) subdirectory of the repository.

I hope this addresses your concern on the low level library documentation.Please let me know your thoughts. Thank you!
