### 2018-05-25 #745 @brainstorm
There seems to be a few function description gaps still left to fill on the http://www.vmtk.org/doc/html/annotated.html link, but overall looks good, what do you think @wouterpotters?
