### 2018-05-25 #745 @wouterpotters
@brainstorm Looks like a great addition to the docs. I vote to accept this manuscript & code for publication.
