### 2018-05-25 #745 @wouterpotters
@kevinmoerman unfortunately I can’t due to technical limitations. I don’t have acces to a physical computer in the next few days and both my iPhone and my Android phone are not allowed (by github) to check that box. Maybe @brainstorm or you could check it?
