### 2018-05-25 #745 @rlizzo
@Kevin-Mattheus-Moerman thank you, I will have some time in a few hours to update the last few functions @brainstorm mentioned. I appreciate the feedback!
