### 2018-05-25 a951470
joss review: completed doxygen documentation as per request
