+++ vtkVmtk/Misc/vtkvmtkMeshVorticity.h
@@ -18,7 +18,7 @@ Version:   $Revision: 1.1 $
      PURPOSE.  See the above copyright notices for more information.
 
 =========================================================================*/
-// .NAME vtkvmtkMeshVorticity - calculates vorticity from velocity gradients in a mesh
+// .NAME vtkvmtkMeshVorticity - Calculates vorticity from velocity gradients in a mesh.
 // .SECTION Description
 // .
 
