### 2018-05-26 #745 @rlizzo
@brainstorm and @Kevin-Mattheus-Moerman, I have updated the documentation at your request. See changes in http://www.vmtk.org/doc/html/annotated.html.
