### 2018-05-26 #745 @rlizzo
Thanks everyone! Archived the version with all the boxes checked on zendoo:

[![DOI](https://zenodo.org/badge/842163.svg)](https://zenodo.org/badge/latestdoi/842163)
